Build export file names from a whole frame number
The frame was passed as a double and written straight to the name, so
past six significant digits it came out in scientific notation: frame
1000000 gave "render.1e+06.tif". Frame numbers are whole, so take one.

+++ lib/djv/App/ExportTool.cpp
@@ -567,7 +567,8 @@ namespace djv
                     const std::string fileName = getExportFileName(
                         options,
                         fileType,
-                        p.exportData->range.start_time().value());
+                        static_cast<int64_t>(
+                            p.exportData->range.start_time().value()));
                     p.exportData->path = ftk::Path(options.dir, fileName);
 
                     // Get the writer.
